Consolidate lace_metadata
We're starting over, so let's use a different versionsing scheme
