+++ lace/lace_metadata/src/utils.rs
@@ -9,9 +9,7 @@ use log::info;
 use rand_xoshiro::Xoshiro256Plus;
 use serde::{Deserialize, Serialize};
 
-use crate::latest::Codebook;
-use crate::versions::v1::DataStore;
-use crate::versions::v2::DatalessState;
+use crate::latest::{Codebook, DataStore, DatalessState};
 use crate::{Error, FileConfig, SerializedType};
 
 fn extenson_from_path<P: AsRef<Path>>(path: &P) -> Result<&str, Error> {
