--- lace/lace_metadata/src/versions/mod.rs
@@ -1,2 +1,0 @@
-pub mod v1;
-pub mod v2;
